requirements.txt
Pillow
numpy
attrs